SBERDOMA-508 refactored comments moved `CommentsList` into `Comments/index`
